requirements.txt
# 后端依赖
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
httpx==0.25.0
redis==4.6.0
qdrant-client==1.5.3
PyJWT==2.8.0
python-dotenv==1.0.0
numpy==1.24.3
networkx==3.1
PyYAML==6.0.1

# 文档处理
PyMuPDF==1.22.5
python-docx==0.8.11
openpyxl==3.1.2
markdown==3.4.4
beautifulsoup4==4.12.2

# 前端
streamlit==1.27.2
pandas==2.0.3
plotly==5.17.0

# 工具和辅助库
tqdm==4.66.1
tenacity==8.2.3
ujson==5.8.0